+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file, in reverse 
 ### Added
 
 - [#26](https://github.com/zendframework/zend-navigation/pull/26) adds:
+  - `Zend\Navigation\View\ViewHelperManagerDelegatorFactory`, which decorates
+    the `ViewHelperManager` service to configure it using
+    `Zend\Navigation\View\HelperConfig`.
   - `ConfigProvider`, which maps the default navigation factory and the
     navigation abstract factory, as well as the navigation view helper.
   - `Module`, which does the same as the above, but for zend-mvc
@@ -14,9 +17,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Deprecated
 
-- [#26](https://github.com/zendframework/zend-navigation/pull/26) deprecates
-  `Zend\Navigation\View\HelperConfig`, as the functionality is now provided
-  by the above additions.
+- Nothing.
 
 ### Removed